server/requirements.txt
# Requirements for mission control server
# pip3 install -r requirements.txt

pyserial
python-socketio
gevent
gevent-websocket
asyncssh
uvicorn
psutil